Added examples
